chore(NA): moving @kbn/utility-types into bazel (#97151)
* chore(NA): moving @kbn/utility-types into bazel

* chore(NA): solve ts config errors
