+++ package.json
@@ -138,7 +138,7 @@
     "@kbn/tinymath": "link:bazel-bin/packages/kbn-tinymath/npm_module",
     "@kbn/ui-framework": "link:packages/kbn-ui-framework",
     "@kbn/ui-shared-deps": "link:packages/kbn-ui-shared-deps",
-    "@kbn/utility-types": "link:packages/kbn-utility-types",
+    "@kbn/utility-types": "link:bazel-bin/packages/kbn-utility-types/npm_module",
     "@kbn/utils": "link:packages/kbn-utils",
     "@loaders.gl/core": "^2.3.1",
     "@loaders.gl/json": "^2.3.1",
