+++ packages/kbn-utility-types/tsd_tests/package.json
@@ -1,0 +1,9 @@
+{
+  "types": "empty.d.ts",
+  "tsd": {
+    "directory": "test_d",
+    "compilerOptions": {
+      "rootDir": "../"
+    }
+  }
+}
