+++ packages/kbn-utility-types/tsd_tests/test_d/public_methods_of.ts
@@ -6,8 +6,9 @@
  * Side Public License, v 1.
  */
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { expectAssignable, expectNotAssignable } from 'tsd';
-import { PublicMethodsOf } from '../index';
+import { PublicMethodsOf } from '../../index';
 
 class Test {
   public name: string = '';
