+++ packages/kbn-utility-types/tsd_tests/test_d/shallow_promise.ts
@@ -6,8 +6,9 @@
  * Side Public License, v 1.
  */
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { expectType } from 'tsd';
-import { ShallowPromise } from '../index';
+import { ShallowPromise } from '../../index';
 
 type P1 = ShallowPromise<string>;
 type P2 = ShallowPromise<ShallowPromise<string>>;
