+++ packages/kbn-utility-types/tsd_tests/test_d/writable.ts
@@ -6,8 +6,9 @@
  * Side Public License, v 1.
  */
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { expectAssignable } from 'tsd';
-import { Writable } from '../index';
+import { Writable } from '../../index';
 
 type WritableArray = Writable<readonly string[]>;
 expectAssignable<WritableArray>(['1']);
